# mentor-marketplace-openapi.yaml
openapi: 3.0.3
info:
  title: Mentor Marketplace Backend API
  version: 1.0.0
  description: |
    OpenAPI specification for a real-time mentor marketplace backend.
    Includes authentication (OTP), user profiles, mentor discovery, booking lifecycle,
    session OTP start/end, and payment operations.
servers:
  - url: https://api.mentor-marketplace.com/v1
    description: Production
  - url: https://staging-api.mentor-marketplace.com/v1
    description: Staging
...
security:
  - bearerAuth: []
paths:
  /auth/otp/send:
    post:
      tags: [Auth]
      summary: Send OTP to phone number
      description: Sends a one-time password to the given phone for login/signup.
      security: []
      operationId: sendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendOtpResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP and issue tokens
      security: []
      operationId: verifyOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyOtpResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /payments:
    get:
      tags: [Payments]
      summary: List user payments
      operationId: listPayments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Payment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPaymentsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /payments/{paymentId}:
    get:
      tags: [Payments]
      summary: Get payment details
      operationId: getPaymentById
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/{paymentId}/refund:
    post:
      tags: [Payments]
      summary: Create refund
      operationId: createRefund
      parameters:
        - $ref: '#/components/parameters/PaymentId'
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRefundRequest'
      responses:
        '201':
          description: Refund created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict with current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    TooManyRequests:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              example: INVALID_REQUEST
            message:
              type: string
              example: Request payload is invalid.
            requestId:
              type: string
              example: req_1A2B3C4D
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          additionalProperties: false
          properties:
            error:
              type: object
              additionalProperties: false
              properties:
                fields:
                  type: array
                  items:
                    $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
        code:
          type: string
          example: REQUIRED
        message:
          type: string
          example: categoryId is required.
    SendOtpRequest:
      type: object
      additionalProperties: false
      required: [phoneNumber, countryCode, purpose]
      properties:
        phoneNumber:
          type: string
          pattern: '^[0-9]{6,15}$'
          example: '9876543210'
        countryCode:
          type: string
          pattern: '^\+[1-9][0-9]{0,3}$'
          example: '+91'
        purpose:
          type: string
          enum: [login]
        deviceId:
          type: string
          maxLength: 128
    SendOtpResponse:
      type: object
      additionalProperties: false
      required: [otpRequestId, expiresAt, resendAfterSeconds, deliveryHint]
      properties:
        otpRequestId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        resendAfterSeconds:
          type: integer
          minimum: 0
        deliveryHint:
          type: string
          enum: [sms, dev_log, none]
          description: >
            sms: Twilio is configured; code should arrive by SMS.
            dev_log: OTP_LOG_CODE is on; code is printed in API logs.
            none: no SMS and no log mode; configure Twilio or OTP_LOG_CODE for local dev.
    VerifyOtpRequest:
      type: object
      additionalProperties: false
      required: [otpRequestId, otpCode]
      properties:
        otpRequestId:
          type: string
          format: uuid
        otpCode:
          type: string
          pattern: '^[0-9]{4,8}$'
        deviceId:
          type: string
          maxLength: 128
    VerifyOtpResponse:
      type: object
      additionalProperties: false
      required: [accessToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
          minimum: 60
        user:
          $ref: '#/components/schemas/UserSummary'
    UserSummary:
      type: object
      additionalProperties: false
      required: [id, phoneNumber, roles]
      properties:
        id:
          type: string
          format: uuid
        phoneNumber:
          type: string
        roles:
          type: array
          minItems: 1
          items:
            type: string
            enum: [mentee, mentor]
...
    ListPaymentsResponse:
      type: object
      additionalProperties: false
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PaymentResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Payment'
    Payment:
      type: object
      additionalProperties: false
      required:
        [paymentId, bookingId, payerUserId, amount, currency, status, provider, providerReference, commission, createdAt, updatedAt]
      properties:
        paymentId:
          type: string
          format: uuid
        bookingId:
          type: string
          format: uuid
        payerUserId:
          type: string
          format: uuid
        amount:
          type: integer
          minimum: 0
          description: Gross amount in minor unit.
        currency:
          type: string
          enum: [INR]
        status:
          type: string
          enum: [pending, authorized, captured, failed, refunded, partially_refunded]
        provider:
          type: string
          example: razorpay
        providerReference:
          type: string
        commission:
          $ref: '#/components/schemas/CommissionBreakdown'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRefundRequest:
      type: object
      additionalProperties: false
      required: [amount, reason]
      properties:
        amount:
          type: integer
          minimum: 1
          description: Amount in minor unit.
        reason:
          type: string
          enum: [mentor_no_show, duplicate_charge, user_cancellation, quality_issue, other]
        reasonNote:
          type: string
          maxLength: 300
    RefundResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Refund'
    Refund:
      type: object
      additionalProperties: false
      required: [refundId, paymentId, amount, currency, reason, status, createdAt]
      properties:
        refundId:
          type: string
          format: uuid
        paymentId:
          type: string
          format: uuid
        amount:
          type: integer
        currency:
          type: string
          enum: [INR]
        reason:
          type: string
        status:
          type: string
          enum: [pending, succeeded, failed]
        createdAt:
          type: string
          format: date-time
...
    CommissionBreakdown:
      type: object
      additionalProperties: false
      required: [takeRatePercent, platformAmount, mentorAmount]
      properties:
        takeRatePercent:
          type: number
          format: float
          minimum: 0
          maximum: 100
          example: 20
        platformAmount:
          type: integer
          minimum: 0
        mentorAmount:
          type: integer
          minimum: 0
...
    Pagination:
      type: object
      additionalProperties: false
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer